feat: a bit of structure
