+++ tab-group/tab-group.js
@@ -1,90 +1,4 @@
-/**
- * Focuses without scroll and then smoothly scrolls horizontally to the tab.
- * The two operations are separated by a frame.
- * @param {HTMLElement} element The element to focus and scroll to.
- * @returns {void}
- */
-const focusAndScroll = (element) => {
-  requestAnimationFrame(() => {
-    element.focus({ preventScroll: true });
-
-    requestAnimationFrame(() => {
-      // Smoothly scroll to the element, aligning it with the start of the
-      // container to ensure the tab is fully visible.
-      element.scrollIntoView({
-        behavior: "smooth",
-        inline: "start",
-      });
-    });
-  });
-};
-
-/**
- * Implements the manual keyboard navigation for the tab group.
- * @param {KeyboardEvent} evt The `keydown` event triggered on a tab's button.
- * @param {Number} index The index of the tab in the collection.
- * @returns {void}
- */
-const navigateTabs = (evt, index) => {
-  // Keyboard navigation is handled but tabbing is left to the browser
-  if (evt.key === "Tab") {
-    return;
-  }
-
-  evt.preventDefault();
-  evt.stopPropagation();
-
-  const containingTabList = evt.currentTarget.parentElement?.parentElement;
-
-  if (containingTabList?.role !== "tablist") {
-    throw new Error("Broken structure, unable to proceed");
-  }
-
-  /**
-   * @type {NodeListOf<HTMLButtonElement>} tabButtons The buttons in the tab list.
-   */
-  const tabButtons = containingTabList.querySelectorAll('[role="tab"] button');
-
-  // Useful names for the indices
-  const first = 0,
-    last = tabButtons.length - 1,
-    next = index + 1,
-    previous = index - 1;
-
-  // With the default prevented, the "click" equivalent needs to be triggered
-  if (evt.key === "Enter" || evt.key === " " || evt.code === "Space") {
-    evt.currentTarget.click();
-  }
-  // Otherwise handle the navigation:
-  // Going to the first tab
-  else if (evt.key === "Home" || (evt.key === "ArrowLeft" && evt.metaKey)) {
-    focusAndScroll(tabButtons[first]);
-  }
-  // Or going to the last tab
-  else if (evt.key === "End" || (evt.key === "ArrowRight" && evt.metaKey)) {
-    focusAndScroll(tabButtons[last]);
-  }
-  // Moving to the right (next tab)
-  else if (evt.key === "ArrowRight") {
-    if (next < tabButtons.length) {
-      focusAndScroll(tabButtons[next]);
-    }
-    // Loop back to the start if the focus is at the end
-    else {
-      focusAndScroll(tabButtons[first]);
-    }
-  }
-  // Moving to the left (previous tab)
-  else if (evt.key === "ArrowLeft") {
-    if (index > first) {
-      focusAndScroll(tabButtons[previous]);
-    }
-    // Loop around to the end if the focus is at the start
-    else {
-      focusAndScroll(tabButtons[last]);
-    }
-  }
-};
+import { focusAndScroll, navigateTabs } from "./tab-behaviors";
 
 class TabGroup extends HTMLElement {
   static get observedAttributes() {
@@ -103,57 +17,66 @@ class TabGroup extends HTMLElement {
   }
 
   #selectedTabIndex = 0;
+  #mounted = false;
 
   connectedCallback() {
-    const base = document
-      .getElementById("tab-group-structure")
-      .content.cloneNode(true);
-    const tabList = base.querySelector('[role="tablist"]');
+    if (!this.#mounted) {
+      const base = document
+        .getElementById("tab-group-structure")
+        .content.cloneNode(true);
+      const tabList = base.querySelector('[role="tablist"]');
 
-    this.querySelectorAll("details").forEach((details, index) => {
-      const detailsClone = details.cloneNode(true);
+      this.querySelectorAll("details").forEach((details, index) => {
+        const detailsClone = details.cloneNode(true);
 
-      const summary = detailsClone.querySelector("summary");
+        const summary = detailsClone.querySelector("summary");
 
-      const title = summary.textContent;
+        const title = summary.textContent;
 
-      detailsClone.removeChild(summary);
+        detailsClone.removeChild(summary);
 
-      const pairId = TabGroup.#nextId;
+        const pairId = TabGroup.#nextId;
 
-      const tab = document.createElement("li");
-      tab.role = "tab";
-      tab.id = `tab-${pairId}`;
-      tab.ariaSelected = index === this.#selectedTabIndex;
-      // Non-reflected attribute
-      tab.setAttribute("aria-controls", `panel-${pairId}`);
+        const tab = document.createElement("li");
+        tab.role = "tab";
+        tab.id = `tab-${pairId}`;
+        tab.ariaSelected = index === this.#selectedTabIndex;
+        // Non-reflected attribute
+        tab.setAttribute("aria-controls", `panel-${pairId}`);
 
-      const tabButton = document.createElement("button");
-      tabButton.type = "button";
-      tabButton.textContent = title;
-      tabButton.tabIndex = index === this.#selectedTabIndex ? 0 : -1;
-      tabButton.addEventListener("keydown", (evt) => navigateTabs(evt, index), {
-        passive: false,
+        const tabButton = document.createElement("button");
+        tabButton.type = "button";
+        tabButton.textContent = title;
+        tabButton.tabIndex = index === this.#selectedTabIndex ? 0 : -1;
+        tabButton.addEventListener(
+          "keydown",
+          (evt) => navigateTabs(evt, index),
+          {
+            passive: false,
+          }
+        );
+        tabButton.addEventListener("click", () => {
+          // Private setter?
+          this.#selectedTabIndex = index;
+          this.#update();
+        });
+        tab.appendChild(tabButton);
+        tabList.appendChild(tab);
+
+        const panel = document.createElement("section");
+        panel.role = "tabpanel";
+        panel.id = `panel-${pairId}`;
+        panel.hidden = index !== 0;
+        // Non-reflected attribute
+        panel.setAttribute("aria-labelledby", `tab-${pairId}`);
+        panel.innerHTML = detailsClone.innerHTML;
+        base.appendChild(panel);
       });
-      tabButton.addEventListener("click", () => {
-        // Private setter?
-        this.#selectedTabIndex = index;
-        this.#update();
-      });
-      tab.appendChild(tabButton);
-      tabList.appendChild(tab);
 
-      const panel = document.createElement("section");
-      panel.role = "tabpanel";
-      panel.id = `panel-${pairId}`;
-      panel.hidden = index !== 0;
-      // Non-reflected attribute
-      panel.setAttribute("aria-labelledby", `tab-${pairId}`);
-      panel.innerHTML = detailsClone.innerHTML;
-      base.appendChild(panel);
-    });
-
-    this.replaceChildren(base);
+      this.replaceChildren(base);
+      // The transformation of the content is unidirectional
+      this.#mounted = true;
+    }
 
     this.#update();
   }
@@ -195,12 +118,24 @@ class TabGroup extends HTMLElement {
     if (v === "underline") {
       return "underline";
     } else {
+      // This is the default, also when unsupported values are passed.
       return "pill";
     }
   }
 
   get title() {
     return this.getAttribute("title");
+  }
+
+  /**
+   * @param {String} v
+   */
+  set variant(v) {
+    this.setAttribute("variant", v);
+  }
+
+  set title(t) {
+    this.setAttribute("title", t);
   }
 }
